structured and refactored the code in some new files
