+++ lib/chore_list.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'chore.dart';
+import 'api.dart';
 
 // handles all changes to the list with chores
 class ChoreList extends ChangeNotifier {
